refactor: extract CLI handlers into Tricorder.CLI
- Move `showStatus`, `showLog`, `showSource` from `Tricorder` into new `Tricorder.CLI`
- Rename `Tricorder.Render` → `Tricorder.CLI.Render`
- Move `waitForDaemon` to `Tricorder.Daemon`
- Inline small `start`/`stop`/`ui` helpers back into the top-level `run` dispatch
